pyproject.toml
[project]
name = "devboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.136.1",
    "psycopg2>=2.9.12",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.14.0",
    "pyjwt>=2.12.1",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
]
